+++ client/app/admin/admin.component.js
@@ -3,6 +3,7 @@
 import angular from 'angular';
 import uiRouter from 'angular-ui-router';
 import routes from './admin.routes';
+import tabs from 'angular-ui-bootstrap/src/tabs';
 import UserManager from '../../components/usermanager/usermanager.component';
 import ScheduleManager from '../../components/schedulemanager/schedulemanager.component';
 import AnnouncementManager from '../../components/announcementmanager/announcementmanager.component';
@@ -11,7 +12,7 @@ export class AdminController {
   /*@ngInject*/
 }
 
-export default angular.module('shyApp.admin', [uiRouter, UserManager, ScheduleManager, AnnouncementManager])
+export default angular.module('shyApp.admin', [uiRouter, tabs, UserManager, ScheduleManager, AnnouncementManager])
   .config(routes)
   .component('admin', {
     template: require('./admin.pug'),