+++ client/app/classes/classes.component.js
@@ -2,6 +2,9 @@
 import angular from 'angular';
 import uiRouter from 'angular-ui-router';
 import routes from './classes.routes';
+import amPM from '../../components/ampm/ampm.filter';
+import daytodate from '../../components/daytodate/daytodate.filter';
+import htmlid from '../../components/htmlid/htmlid.filter';
 
 export class ClassesController {
   /*@ngInject*/
@@ -23,7 +26,7 @@ export class ClassesController {
   }
 }
 
-export default angular.module('shyApp.classes', [uiRouter])
+export default angular.module('shyApp.classes', [uiRouter, amPM, daytodate, htmlid])
   .config(routes)
   .component('classes', {
     template: require('./classes.pug'),