+++ client/app/locations/locations.component.js
@@ -2,6 +2,8 @@
 import angular from 'angular';
 import routes from './locations.routes';
 import uiRouter from 'angular-ui-router';
+import htmlid from '../../components/htmlid/htmlid.filter';
+import trustedurl from '../../components/trustedurl/trustedurl.filter';
 
 export class LocationsController {
   /*@ngInject*/
@@ -18,7 +20,7 @@ export class LocationsController {
   }
 }
 
-export default angular.module('shyApp.locations', [uiRouter])
+export default angular.module('shyApp.locations', [uiRouter, htmlid, trustedurl])
   .config(routes)
   .component('locations', {
     template: require('./locations.pug'),