+++ client/app/teachers/teachers.component.js
@@ -2,6 +2,8 @@
 import angular from 'angular';
 import routes from './teachers.routes';
 import uiRouter from 'angular-ui-router';
+import htmlid from '../../components/htmlid/htmlid.filter';
+import nosubs from '../../components/nosubs/nosubs.filter';
 
 export class TeachersController {
   /*@ngInject*/
@@ -19,7 +21,7 @@ export class TeachersController {
   }
 }
 
-export default angular.module('shyApp.teachers', [uiRouter])
+export default angular.module('shyApp.teachers', [uiRouter, htmlid, nosubs])
   .config(routes)
   .component('teachers', {
     template: require('./teachers.pug'),