+++ client/components/navbar/navbar.component.js
@@ -1,7 +1,8 @@
 /* eslint no-sync:0 */
 'use strict';
-
 import angular from 'angular';
+import collapse from 'angular-ui-bootstrap/src/collapse';
+import dropdown from 'angular-ui-bootstrap/src/dropdown/index-nocss.js';
 
 export class NavbarController {
   /*@ngInject*/
@@ -96,7 +97,7 @@ export class ContactModalController {
   }
 }
 
-export default angular.module('shyApp.navbar', [])
+export default angular.module('shyApp.navbar', [collapse, dropdown])
   .component('navbar', {
     template: require('./navbar.pug'),
     controller: NavbarController